create the purchaser model
our purchaser is a very simple model, since all we have to go off of is
a name. this raises all sorts of problems if two purchasers have the
same name but since that's the only information we're getting from our
acquisition we'll have to assume they required purchaser names to be
unique.