create the merchant model
the data we have for merchants include both their name and their
address. it's possible for two merchants to have the same name and
different addresses (perhaps two independently owned franchises) or for
two merchants to have the same address (located in the same office
building), so we enforce unqiueness on the combination of the two. as a
backup, we also enforce uniqueness on the index at the database level.