add the item model
I would love to use a better than 'Item' here but it's best not to guess
out of context. 'offer' and 'product' both seem a bit too exclusionary
and we only have a small set of exmaple data to go off of.

Similar to the Merchant situation, the only way to determine the
uniqueness of an Item is by looking at several columns. Here we have to
look at the description, price, and merchant id. It's possible for a
merchant to offer the same description with multiple prices (discounts
for membership, etc.) or for multiple merchants to offer an item with
the same description (competing businesses).