make everything a bit more aesthetically pleasing
I know it's not required so I won't consider the extra work here in my
total time spent on the solution. However, I couldn't stand looking at
the application. I also added a but more information.

note: some tests had to be updated because the content on the page was
updated.